combine ding and email in notify
